% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_algorithms.R
\name{inv_swap}
\alias{inv_swap}
\title{Nonnative species swap taxonomic beta diversity change null model algorithm}
\usage{
inv_swap(mod_com, his_com, region_bridge, const = "reg.all")
}
\arguments{
\item{mod_com}{Community matrix containing all species (native+nonnative)
with columns for species and rows for site. Must have the same row names as
mod_com.}

\item{his_com}{Community matrix containing only native species with
columns for species and rows for site. Must have the same row names as
mod_com. "Pseudo-historical" data}

\item{region_bridge}{data.frame with column COMID that matches row names
of com, and HUC_8 which define that site's region, and HUC_8 which defines
a site's subbasin (used for native vs nonnative identification)}

\item{const}{character ("reg.all","reg.cont", or "reg.freq"). Defines
constraints to null model randomization. See details for more information}
}
\value{
community matrix with identical dimensions and labels as mod_com,
but random distribution of nonnative species occurrences
}
\description{
Randomizes the occurrences of nonnative species in a
contemporary community data set, for use in taxonomic beta diversity change
null models.
}
\details{
Null model algorithm based on Leprieur et al., (2008), that randomizes the
occurrences of nonnative species to test if directional  beta diversity
change is more extreme then expected by random distributions of nonnative
species. This algorithm should only be used in pseudo-historic approaches,
where full community data is compared to community data with nonnative species
withheld as pseudo-historical community data.

This null model maintains an equiprobable total of columns by randomizing the
spatial distribution of nonnative species, maintaining their occurrence
frequency, but allowing communities to be unconstrained in the number of
nonnative species they receive. Leprieur et al., (2008) provides detailed
arguments for the ecological validity of this equiprobable null model in
the context of invasion-driven Δβ. In short, the authors argue that the
model fits the expectation that nonnative species differ in their
colonization ability and propagule pressure, most communities are
susceptible to invasions, and that communities are rarely saturated.

While this algorithm contains the occurrence frequency of each
nonnative species to observed frequencies, this function allows
users to specify additional constraints. "reg.all" allows for a nonnative
species to be placed into any region (HUC_2) and does not maintain regional
occurrence frequencies; "reg.cont" restricts nonnative species into regions
where they have been observed, nonnative but does not maintain regional
occurrence frequencies; and "reg.freq" restrict nonnative species
to regions where they have occurred and maintains regional
occurrence frequencies.
}
