% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_batch_functions.R
\name{kernel_beta_batch}
\alias{kernel_beta_batch}
\title{Batch estimate kernel density based  functional beta diversity and lcbd}
\usage{
kernel_beta_batch(com, trait_hyp, cores = 1, abund = F)
}
\arguments{
\item{com}{Community matrix with columns for species and rows for site}

\item{trait_hyp}{Trait hypervolume created using BAT::hyper.build}

\item{cores}{Number of cores to be used in parallel processing. If = 0 all available cores are used. Beware that multicore for Windows is not optimized yet and it often takes longer than single core.}

\item{abund}{A boolean (T/F) indicating whether abundance data should be used as weights in hypervolume construction. Only works if method.hv = "gaussian".}
}
\value{
Named list containing pairwise beta diversity output list from
BAT::kernel.beta and a data.frame containing LCBD values for total,
replacement, and richness difference components
}
\description{
Estimates functional beta diversity and local contribution to
beta diversity (LCBD) using kernel density hypervolumes. Designed to run
within the parallel function
}
