% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_batch_functions.R
\name{phy_beta_batch}
\alias{phy_beta_batch}
\title{Batch estimate phylogenetic beta diversity and lcbd}
\usage{
phy_beta_batch(com, tree, abund = F)
}
\arguments{
\item{com}{Community matrix with columns for species and rows for site}

\item{tree}{A phylo or hclust object (used only for PD or FD) or alternatively a species x traits matrix or data.frame to build a functional tree.}

\item{abund}{A boolean (T/F) indicating whether abundance data should be used or converted to incidence before analysis.}
}
\value{
Named list containing pairwise beta diversity output list from
BAT::beta and a data.frame containing LCBD values for total,
replacement, and richness difference components
}
\description{
Estimates phylogenetic beta diversity and local contribution to
beta diversity (LCBD). Designed to run within the parallel function
}
