% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_effect_size_function.R
\name{ses_fun}
\alias{ses_fun}
\title{Null model standardized effect size and diagnostic function}
\usage{
ses_fun(obs, null_list, emp_padding = 0, diag = T, mc.cores = 1)
}
\arguments{
\item{obs}{vector, dist, or matrix containing observed diversity values}

\item{null_list}{list including null iterations of diversity metric. Must be
in same format as obs}

\item{emp_padding}{value used to pad empirical p values to prevent inf
effect sizes. If p value = zero, effect is infinite. This can be corrected
by adding 1 to the numerator and denominator of p value}

\item{diag}{Logical. If true, returns measures of skew, and kurtosis
for each null distribution. Used to assess normality of null distribution.}

\item{mc.cores}{numerical. Number of cores for parallel processing}
}
\value{
List containing:means of null distributions (null_mean), standard
deviation of null distribution (null_sd), z-score based standardized effect
sizes (ses), empirical_pvalues, empirical_pvalue-based effect size (ES),
skewness of null distribution (skew), and kurtosis of null
distribution (kurt)
}
\description{
Estimates null model standardized effects sizes and returns
diagnostics statistics for null distributions.
}
\details{
his function is flexible and takes a range of input formats such as
data.frames, vectors, matrices, and distance objects, and maintains this
format in the exported values. The function estimates standardize effect
sizes in the traditional z score method (SES). Additionally empirical
p-values based effect sizes (ES) are calculated. Empirical p-values describe
the proportion of the null model that is more extreme than the observed
values and p-value based effect sizes (ES) are calculated. Finally the
function reports optional diagnostic metric to assess if null distributions
are symmetrical and normal. asymmetrical null distributions should be
assessed using empirical p-value based effect sizes rather than z-score
based SES. See Botta-Dukát (2018) for more information on selecting SES or
p-value based ES.
}
