% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_batch_functions.R
\name{tax_beta_batch}
\alias{tax_beta_batch}
\title{Batch estimate taxonomic beta diversity and lcbd}
\usage{
tax_beta_batch(com, abund = F)
}
\arguments{
\item{com}{Community matrix with columns for species and rows for site}

\item{abund}{A boolean (T/F) indicating whether abundance data should be used or converted to incidence before analysis.}
}
\value{
Named list containing pairwise beta diversity output list from
BAT::beta and a data.frame containing LCBD values for total,
replacement, and richness difference components
}
\description{
Estimates taxonomic beta diversity and local contribution to
beta diversity (LCBD). Designed to run within the parallel function
}
