% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_algorithms.R
\name{taxa_swap_region}
\alias{taxa_swap_region}
\title{Regional constrained taxa-swap null model algorithm}
\usage{
taxa_swap_region(com, region_bridge)
}
\arguments{
\item{com}{Community matrix with columns for species and rows for site}

\item{region_bridge}{data.frame with column COMID that matches row names
of com, and HUC_2 which define that site's region.}
}
\value{
randomly shuffled community matrix with identical labels and
dimensions to input matrix
}
\description{
Random shuffles species labels in community matrix, constraining
swaps to some defined regional species pool (HUC_2)
}
