% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_algorithms.R
\name{trait_swap}
\alias{trait_swap}
\title{Functional trait taxa-swap null model algorithm}
\usage{
trait_swap(trait)
}
\arguments{
\item{trait}{matrix with columns for functional traits, and rows for species}
}
\value{
randomly shuffled trait matrix with identical labels and
dimensions to input matrix
}
\description{
Random shuffles species labels within a functional trait matrix
}
