% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_algorithms.R
\name{tree_swap}
\alias{tree_swap}
\title{Phylogenetic tree taxa-swap null model algorithm}
\usage{
tree_swap(tree)
}
\arguments{
\item{tree}{phylogenetic tree}
}
\value{
phylogenetic tree with randomized tip labes
}
\description{
Random shuffles species labels within a phylognetic tree
}
